Wrap with CBC for symmetric, not GCM
GCM is inferior to CBC here for two reasons.

1. GCM is much more wholly broken in the case that IV reuse happens for
a key.
2. GCM has a smaller IV (96 bits vs 128 bits) than CBC.

Or, to put it in terms of NIST's recommendations, you shouldn't use GCM
for more than 2^32 encryptions with a single key,
https://csrc.nist.gov/publications/detail/sp/800-38d/final (section 8.3).

Which the current setup encourage.

By using CBC instead, the birthday paradox becomes a lot less likely to
be hit (16 more bits of entropy in the IV), and the cost of hitting it
is much less drastic.

Also, given we're generally encrypting random data here (e.g. crypto
keys), the cost of IV reuse is basically zero if you use CBC, but is
high with GCM (because for an attacker, if you manage to birthday
paradox it, you'd be able to in the worst case retrieve the whole key,
whereas with CBC, you would only be able to figure out that two keys
start with the same prefix.